Add openSearchForm() utility function
Import openSearchForm in hotkeys.js
